+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Display current Slider value when hovered.
+- Hold Ctrl to snap Slider value to 0.01 increments.
 
 ## [1.1.1] - 2022-07-30
 ### Fixed